Added Node Generation Code
Changed instance to supposed smallest instance, 0o50t100s1p125

Added code to generate nodes. A node is a courier (off-time, restaurant, time) triple, with nodes for the same (off-time, restaurant) pair differing by the given nodeTimeInterval.

We don't consider an (off-time, restaurant) pair if it is not possible for a courier  with the given off-time to get to the restaurant in time to deliver an order, before their off-time.