Generalising groups and other small changes
A courier group has a fixed off time for all couriers in the group. It is defined by some group identifier, followed by a list, where the first entry is a list of all the couriers in that group, and the second is the off time of all couriers in the group.
Changes made up to and including the creation of the untimed arcs.